+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.1.10 - ????
+## 0.2.0 - ????
 
 ### Features
 
@@ -11,7 +11,12 @@
 -   Add support for permutation variable importance in the GBT learner with the
     `compute_permutation_variable_importance` parameter.
 -   Support for tf.int8 and tf.int16 values.
--   Support for distributed gradient boosted trees learning.
+-   Support for distributed gradient boosted trees learning using the
+    ParameterServerStrategy distribution strategy.
+-   Support for training from dataset stored on disk in CSV and RecordIO format
+    (instead of creating a tensorflow dataset). This option is currently more
+    efficient for distributed training (until the ParameterServerStrategy
+    support per-worker datasets).
 -   Add `max_vocab_count` argument to the model constructor. The existing
     `max_vocab_count` argument in `FeatureUsage` objects take precedence.